vm_docker/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
scikit-learn==1.2.2
joblib==1.2.0
python-multipart==0.0.6